Updated Elasticsearch read store tests
