+++ Source/EventFlow.MsSql.Tests/EventFlow.MsSql.Tests.csproj
@@ -73,13 +73,13 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="IntegrationTestsForEventStore\MsSqlEventStoreTests.cs" />
-    <Compile Include="IntegrationTestsForReadStore\MsSqlReadModelStoreTests.cs" />
-    <Compile Include="IntegrationTestsForReadStore\QueryHandlers\MsSqlThingyGetMessagesQueryHandler.cs" />
-    <Compile Include="IntegrationTestsForReadStore\QueryHandlers\MsSqlThingyGetQueryHandler.cs" />
+    <Compile Include="IntegrationTests\EventStores\MsSqlEventStoreTests.cs" />
+    <Compile Include="IntegrationTests\ReadStores\MsSqlReadModelStoreTests.cs" />
+    <Compile Include="IntegrationTests\ReadStores\QueryHandlers\MsSqlThingyGetMessagesQueryHandler.cs" />
+    <Compile Include="IntegrationTests\ReadStores\QueryHandlers\MsSqlThingyGetQueryHandler.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="IntegrationTestsForReadStore\ReadModels\MsSqlThingyMessageReadModel.cs" />
-    <Compile Include="IntegrationTestsForReadStore\ReadModels\MsSqlThingyReadModel.cs" />
+    <Compile Include="IntegrationTests\ReadStores\ReadModels\MsSqlThingyMessageReadModel.cs" />
+    <Compile Include="IntegrationTests\ReadStores\ReadModels\MsSqlThingyReadModel.cs" />
     <Compile Include="UnitTests\ReadModels\ReadModelSqlGeneratorTests.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -112,10 +112,10 @@
     <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="IntegrationTestsForReadStore\Scripts\0001 - Create table ReadModel-ThingyAggregate.sql" />
+    <EmbeddedResource Include="IntegrationTests\ReadStores\Scripts\0001 - Create table ReadModel-ThingyAggregate.sql" />
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="IntegrationTestsForReadStore\Scripts\0002 - Create table ReadModel-ThingyMessage.sql" />
+    <EmbeddedResource Include="IntegrationTests\ReadStores\Scripts\0002 - Create table ReadModel-ThingyMessage.sql" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
