+++ Source/EventFlow.MsSql.Tests/UnitTests/ReadModels/ReadModelSqlGeneratorTests.cs
@@ -22,8 +22,7 @@
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 // 
 using System.ComponentModel.DataAnnotations.Schema;
-using EventFlow.MsSql.Tests.IntegrationTestsForReadStore;
-using EventFlow.MsSql.Tests.IntegrationTestsForReadStore.ReadModels;
+using EventFlow.MsSql.Tests.IntegrationTests.ReadStores.ReadModels;
 using EventFlow.ReadStores.MsSql;
 using EventFlow.TestHelpers;
 using FluentAssertions;
