+++ Source/EventFlow.ReadStores.Elasticsearch.Tests/EventFlow.ReadStores.Elasticsearch.Tests.csproj
@@ -67,9 +67,11 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Integration\ElasticsearchThingyReadModel.cs" />
-    <Compile Include="Integration\ElasticsearchReadModelStoreTests.cs" />
-    <Compile Include="Integration\ElasticsearchThingyGetQueryHandler.cs" />
+    <Compile Include="IntegrationTests\QueryHandlers\ElasticsearchThingyGetMessagesQueryHandler.cs" />
+    <Compile Include="IntegrationTests\ReadModels\ElasticsearchThingyReadModel.cs" />
+    <Compile Include="IntegrationTests\ElasticsearchReadModelStoreTests.cs" />
+    <Compile Include="IntegrationTests\QueryHandlers\ElasticsearchThingyGetQueryHandler.cs" />
+    <Compile Include="IntegrationTests\ReadModels\ElasticsearchThingyMessageReadModel.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
