+++ Source/EventFlow.ReadStores.Elasticsearch.Tests/IntegrationTests/QueryHandlers/ElasticsearchThingyGetQueryHandler.cs
@@ -25,11 +25,12 @@
 using System.Threading;
 using System.Threading.Tasks;
 using EventFlow.Queries;
+using EventFlow.ReadStores.Elasticsearch.Tests.IntegrationTests.ReadModels;
 using EventFlow.TestHelpers.Aggregates;
 using EventFlow.TestHelpers.Aggregates.Queries;
 using Nest;
 
-namespace EventFlow.ReadStores.Elasticsearch.Tests.Integration
+namespace EventFlow.ReadStores.Elasticsearch.Tests.IntegrationTests.QueryHandlers
 {
     public class ElasticsearchThingyGetQueryHandler : IQueryHandler<ThingyGetQuery, Thingy>
     {
